refactor bigger tests into files
